Move ballot viewer code into ballot screen file
